feat: Add sign-in functionality using next-auth/react
This commit adds sign-in functionality to the AuthForm component using the `next-auth/react` library. The `signIn` function is imported and used to authenticate the user with the provided credentials. If the authentication is successful, a success toast notification is displayed. If there is an error, an error toast notification is displayed. The loading state is also managed to show a loading spinner during the authentication process.

Note: This commit message does not include any issue references, tags, or author names.