feat: Add refresh token expiration field to Account model
This commit adds a new field `refresh_token_expires_in` to the Account model in the Prisma schema. This field will store the expiration time of the refresh token in seconds. This change is necessary to enhance the security and manageability of the authentication system.

Note: This commit message does not include any issue references, tags, or author names.