refactor: Add refresh_token_expires_in field to Account model
This commit adds the refresh_token_expires_in field to the Account model in the prisma/schema.prisma file. The refresh_token_expires_in field is of type Int and represents the expiration time of the refresh token in seconds. This addition enhances the functionality of the application by providing a way to track the expiration of refresh tokens.

Note: This commit message does not include any issue references, tags, or author names.