feat: Add lodash and pusher-js dependencies
This commit adds the "lodash" dependency with a version of "^4.17.21" and the "pusher-js" dependency with a version of "^8.4.0-rc2". These dependencies are added to the package.json file, enabling the usage of lodash utility functions and the Pusher JavaScript library for real-time communication. This addition enhances the functionality of the application by providing additional utility functions and enabling real-time updates.

Note: This commit message does not include any issue references, tags, or author names.